models/aprobaciones_neu_prophet_2026/requirements.txt
pandas
numpy
neuralprophet
scikit-learn
pyyaml
matplotlib
pytest
requests
